Rename and move NotificationsModal and SettingsModal to remove references to modal
