+++ apps/web/src/components/Notifications/CollectionLink.tsx
