+++ apps/web/src/components/Notifications/NotificationUserList/FollowedBackList.tsx
@@ -2,8 +2,8 @@ import { useCallback } from 'react';
 import { usePaginationFragment } from 'react-relay';
 import { graphql } from 'relay-runtime';
 
-import { USERS_PER_PAGE } from '~/components/NotificationsModal/constants';
-import { PureUserList } from '~/components/NotificationsModal/NotificationUserList/PureUserList';
+import { USERS_PER_PAGE } from '~/components/Notifications/constants';
+import { PureUserList } from '~/components/Notifications/NotificationUserList/PureUserList';
 import { FollowedBackListFragment$key } from '~/generated/FollowedBackListFragment.graphql';
 
 type FollowedBackListProps = {
