+++ apps/web/src/components/Notifications/notifications/SomeoneCommentedOnYourFeedEvent.tsx
@@ -6,7 +6,7 @@ import styled from 'styled-components';
 import { VStack } from '~/components/core/Spacer/Stack';
 import { BaseM } from '~/components/core/Text/Text';
 import HoverCardOnUsername from '~/components/HoverCard/HoverCardOnUsername';
-import { CollectionLink } from '~/components/NotificationsModal/CollectionLink';
+import { CollectionLink } from '~/components/Notifications/CollectionLink';
 import { SomeoneCommentedOnYourFeedEventFragment$key } from '~/generated/SomeoneCommentedOnYourFeedEventFragment.graphql';
 import { SomeoneCommentedOnYourFeedEventQueryFragment$key } from '~/generated/SomeoneCommentedOnYourFeedEventQueryFragment.graphql';
 import unescape from '~/utils/unescape';
